Move modules out of domain_model up to top-level of package
