+++ src/vocalpy/sequence.py
