+++ src/vocalpy/spectrogram_parameters.py
